Added server post requests
Changes I made were:
Config.py: A SECRET_KEY was added. This needs to be changed to a more secure solution later.
Model.py: another __repr__ function was added so that when user.query.all() is called it returns both variables
Forms.py: added the new file that creates the wtforms. The two forms it creates are the Createpost and Createlogin forms. These forms recreate the forms on the login page and then create a post page and hence use the same fields in those pages. All fields require validators.
Some more added validation may be required on the image file to ensure it’s not to big and that only one image is being added when the form is submitted.
For the templates
createResquest.html: wtforms post request was added with the action to “/submit”. The checkboxs have not had the wtforms formatting added. This must be done later to do this add another field to the ‘Createpost’ class in the forms.py page. There are good tutorials online.
Login: wtforms post request was added with the actions ‘/submitlogin’ and ‘/createuser’. Switching between the pages still works as before, the js validation code was removed as wtforms validates instead. Form layout is the same. Errors have been implement for the login so that if a username doesn’t exist or the password is wrong it says so.
-	Can add validation to creating a username to have a certain number of characters, same with password.
-	Buttons for submit don’t have the bootstrap look, that needs to be fixed.
Routes.py
Three different submit routes were made.
‘/submit’: submits the create a post page, validates and redirects you to the feed page so hopefully you can see what you uploaded. (that still needs to be implemented)
‘/submitelogin’: submits the login form and checks the username and password are in the system (done by the two functions find_user and find_password). Should redirect you to the posts page if all is correct. However, this function needs to be tested as there may be an error in submitting the password.
‘createuser’: submits a new user, it should add them to the database however this is not implemented yet. Redirects you to the loginform after post.